dz2/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.9
aiohttp==3.11.11
requests==2.32.3
aiohttp==3.11.11
openfoodfacts==2.5.0
